sysloghistorydialog
