+++ src/renderer/src/components/dashboard/EventLog.jsx
@@ -7,119 +7,140 @@ import { useDispatch, useSelector } from 'react-redux'
 import {
   dashboardSelector,
   requestHistoryData,
-  showCustomTableData,
-  updateCustomGraphData
+  showCustomTableData
 } from '../../features/dashboardSlice'
-
 import { Button } from 'antd'
 import { SyncOutlined } from '@ant-design/icons'
 
+function getRandomInt(min = 1, max = 9) {
+  return Math.floor(Math.random() * (max - min + 1)) + min
+}
 const EventLog = (props) => {
-  const { tableData, data } = props
-
+  const { data } = props
   const { customGraphData } = useSelector(dashboardSelector)
   console.log(customGraphData)
   const dispatch = useDispatch()
 
-  const [series, setSeries] = useState([
-    {
-      name: 'Information',
-      color: '#46b300',
-      data: [0, 1, 2, 3, 4, 5]
-    },
-    {
-      name: 'Warning',
-      color: '#F57F17',
-      data: [0, 1, 2, 3, 4, 5]
-    },
-    {
-      name: 'Critical',
-      color: '#D50000',
-      data: [0, 1, 2, 3, 4, 5]
-    }
-  ])
-  const [options, setOptions] = useState({
-    chart: {
-      type: 'bar',
-      height: 100,
-      toolbar: {
-        show: false
+  const [eventLogData, setEventLogData] = useState({
+    series: [
+      {
+        name: 'Information',
+        color: '#46b300',
+        data: [0, 1, 2, 3, 4, 5]
       },
-      offsetY: -20,
-      offsetX: -5
-    },
-
-    legend: {
-      show: true,
-      showForSingleSeries: true,
-      position: 'top',
-      horizontalAlign: 'center',
-      offsetY: 20
-    },
-    plotOptions: {
-      bar: {
-        borderRadius: 0,
-        columnWidth: '50%'
+      {
+        name: 'Warning',
+        color: '#F57F17',
+        data: [0, 1, 2, 3, 4, 5]
+      },
+      {
+        name: 'Critical',
+        color: '#D50000',
+        data: [0, 1, 2, 3, 4, 5]
       }
-    },
-    dataLabels: {
-      enabled: false
-    },
-    stroke: {
-      width: 2
-    },
+    ],
+    options: {
+      chart: {
+        type: 'bar',
+        height: 100,
+        toolbar: {
+          show: false
+        },
+        offsetY: -20,
+        offsetX: -5
+      },
 
-    grid: {
-      show: true
-    },
-    xaxis: {
-      type: 'category',
-      categories: customGraphData.label,
-      labels: {
-        rotate: -45,
-        rotateAlways: true
-      }
-    },
-    yaxis: {
-      title: {
-        lines: {
-          show: true
+      legend: {
+        show: true,
+        showForSingleSeries: true,
+        position: 'top',
+        horizontalAlign: 'center',
+        offsetY: 20
+      },
+      plotOptions: {
+        bar: {
+          borderRadius: 0,
+          columnWidth: '50%'
         }
-      }
-    },
-    fill: {
-      type: 'solid',
+      },
+      dataLabels: {
+        enabled: false
+      },
+      stroke: {
+        width: 2
+      },
 
-      gradient: {
-        shade: 'lights',
-        type: 'horizontal',
-        shadeIntensity: 1,
-        gradientToColors: undefined,
-        inverseColors: false,
-        opacityFrom: 0.85,
-        opacityTo: 0.85,
-        stops: [0, 50]
+      grid: {
+        show: true
+      },
+      xaxis: {
+        type: 'category',
+        categories: customGraphData.label,
+        labels: {
+          rotate: -45,
+          rotateAlways: true
+        }
+      },
+      yaxis: {
+        title: {
+          lines: {
+            show: true
+          }
+        }
+      },
+      fill: {
+        type: 'solid',
+
+        gradient: {
+          shade: 'lights',
+          type: 'horizontal',
+          shadeIntensity: 1,
+          gradientToColors: undefined,
+          inverseColors: false,
+          opacityFrom: 0.85,
+          opacityTo: 0.85,
+          stops: [0, 50]
+        }
       }
     }
   })
 
   useEffect(() => {
-    dispatch(
+    setTimeout(() => {
       requestHistoryData({
-        type: 'custom',
+        type: 'syslog',
         sourceIP: '',
         ge: '',
         le: ''
       })
-    )
+    }, 3000)
   }, [])
 
-  const onCustomGraphClick = () => {
-    dispatch(showCustomTableData(tableData))
+  useEffect(() => {
+    if (Array.isArray(customGraphData.data) && customGraphData.data.length > 0) {
+      setEventLogData((prev) => ({
+        ...prev,
+        series: [
+          {
+            data: customGraphData.data
+          }
+        ],
+        options: {
+          ...prev.options,
+          xaxis: {
+            categories: customGraphData.label
+          }
+        }
+      }))
+    }
+  }, [customGraphData])
+
+  const onCustomGraphClick = (barIndex) => {
+    let tableData = tableData[barIndex]
+    showCustomTableData(tableData)
   }
 
   const handleRefreshGraph = () => {
-    setSeries(series)
     dispatch(
       requestHistoryData({
         type: 'custom',
@@ -131,7 +152,7 @@ const EventLog = (props) => {
   }
 
   return (
-    <div id="chart">
+    <div>
       <div
         style={{
           padding: '0px 5px',
@@ -139,9 +160,9 @@ const EventLog = (props) => {
           justifyContent: 'space-between'
         }}
       >
-        <div>
+        <span>
           <i>{customGraphData.lastUpdated}</i>
-        </div>
+        </span>
         <Button
           title="Refresh"
           icon={<SyncOutlined />}
@@ -150,12 +171,16 @@ const EventLog = (props) => {
         ></Button>
       </div>
       <ReactApexChart
-        options={options}
-        series={series}
+        options={eventLogData.options}
+        series={eventLogData.series}
         type="bar"
         height={210}
         data={data === null ? {} : data}
-        onClick={onCustomGraphClick}
+        onClick={(e, element) => {
+          if (element.length > 0) {
+            onCustomGraphClick(element[0]._index)
+          }
+        }}
       />
     </div>
   )
