+++ src/renderer/src/components/dialogs/SyslogHistoryDialog/SyslogHistoryDialog.jsx
@@ -3,7 +3,7 @@
 import { Modal, Divider, Table, Input, Button, DatePicker } from 'antd'
 import React, { useState, useEffect } from 'react'
 import PropTypes from 'prop-types'
-import { CloseCircleFilled } from '@ant-design/icons'
+//import { CloseCircleFilled } from '@ant-design/icons'
 // import { Close } from '@material-ui/icons'
 //import { RangePicker } from 'antd'
